LucaSupport::Enc::encrypt() / decrypt() use env for automation

+++ lucarecord/lib/luca_support/enc.rb
@@ -13,8 +13,11 @@ module LucaSupport # :nodoc:
     # TODO: check space in dir string
     # TODO: check if origin exists
     def encrypt(dir, iter: 10000, cleanup: false)
+      passopt = ENV['LUCA_ENC_PASSWORD'] ? "-pass pass:#{ENV['LUCA_ENC_PASSWORD']}" : ''
       Dir.chdir(Pathname(CONST.pjdir) / 'data') do
-        system "tar -czf - #{dir} | openssl enc -e -aes256 -iter #{iter} -out #{dir}.tar.gz"
+        abort "Directory #{dir} not found. exit..." unless Dir.exist?(dir)
+
+        system "tar czf - #{dir} | openssl enc -e -aes256 -iter #{iter} -out #{dir}.tar.gz #{passopt}"
         return if ! cleanup
 
         FileUtils.rm_rf dir
@@ -23,8 +26,11 @@ module LucaSupport # :nodoc:
 
     # TODO: check space in dir string
     def decrypt(dir, iter: 10000)
+      passopt = ENV['LUCA_ENC_PASSWORD'] ? "-pass pass:#{ENV['LUCA_ENC_PASSWORD']}" : ''
       Dir.chdir(Pathname(CONST.pjdir) / 'data') do
-        system "openssl enc -d -aes256 -iter #{iter} -in #{dir}.tar.gz | tar xz"
+        abort "Archive #{dir}.tar.gz not found. exit..." unless File.exist?("#{dir}.tar.gz")
+
+        system "openssl enc -d -aes256 -iter #{iter} -in #{dir}.tar.gz #{passopt} | tar xz"
       end
     end
   end
